+++ strafer-liberator-master/core/src/com/port/Portal.java
@@ -36,8 +36,8 @@ public class Portal extends PortalGun {
 
                 getWorld().addObject(new Effect("kingcrimson", 62, 1), 0, 0);
                 
-                player.setWorldX(player.getWorldX()+(this.getX() + Scroller.scrolledX - player.getX()));
-                player.setWorldY(player.getWorldY()+(this.getY() + Scroller.scrolledY - player.getY()));
+                player.setWorldX((int) (player.getWorldX()+(this.getX() + Scroller.scrolledX - player.getX())));
+                player.setWorldY((int) (player.getWorldY()+(this.getY() + Scroller.scrolledY - player.getY())));
                 
                 player.setLocation(this.getX(), this.getY());
